add public

+++ orm-advanced-prototype/routes/web.php
@@ -12,10 +12,10 @@ Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-Route::get('/article/index',[ ArticleController::class,'index'])->name('article.index');
-Route::get('/article/create',[ ArticleController::class,'create'])->name('article.create');
-Route::post('/article',[ ArticleController::class,'store'])->name('article.store');
-Route::delete('/article/{article}',[ ArticleController::class,'destroy'])->name('article.destroy');
+Route::get('/article/index',[ ArticleController::class,'index'])->middleware('role:admin')->name('article.index');
+Route::get('/article/create',[ ArticleController::class,'create'])->middleware('role:admin')->name('article.create');
+Route::post('/article',[ ArticleController::class,'store'])->middleware('role:admin')->name('article.store');
+Route::delete('/article/{article}',[ ArticleController::class,'destroy'])->middleware('role:admin')->name('article.destroy');
